--- Match Tracker Extension/Assets.xcassets/Complication.complicationset/Contents.json
@@ -1,28 +1,0 @@
-{
-  "assets" : [
-    {
-      "idiom" : "watch",
-      "filename" : "Circular.imageset",
-      "role" : "circular"
-    },
-    {
-      "idiom" : "watch",
-      "filename" : "Extra Large.imageset",
-      "role" : "extra-large"
-    },
-    {
-      "idiom" : "watch",
-      "filename" : "Modular.imageset",
-      "role" : "modular"
-    },
-    {
-      "idiom" : "watch",
-      "filename" : "Utilitarian.imageset",
-      "role" : "utilitarian"
-    }
-  ],
-  "info" : {
-    "version" : 1,
-    "author" : "xcode"
-  }
-}
