--- Match Tracker Extension/Assets.xcassets/Complication.complicationset/Modular.imageset/Contents.json
@@ -1,18 +1,0 @@
-{
-  "images" : [
-    {
-      "idiom" : "watch",
-      "scale" : "2x",
-      "screen-width" : "<=145"
-    },
-    {
-      "idiom" : "watch",
-      "scale" : "2x",
-      "screen-width" : ">145"
-    }
-  ],
-  "info" : {
-    "version" : 1,
-    "author" : "xcode"
-  }
-}
