--- Match Tracker/Base.lproj/Interface.storyboard
@@ -1,69 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder.WatchKit.Storyboard" version="3.0" toolsVersion="14113" targetRuntime="watchKit" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" colorMatched="YES" initialViewController="AgC-eL-Hgc">
-    <device id="watch38" orientation="portrait">
-        <adaptation id="fullscreen"/>
-    </device>
-    <dependencies>
-        <deployment identifier="watchOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="14088"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBWatchKitPlugin" version="14033"/>
-    </dependencies>
-    <scenes>
-        <!--Interface Controller-->
-        <scene sceneID="aou-V4-d1y">
-            <objects>
-                <controller id="AgC-eL-Hgc" customClass="InterfaceController" customModule="Match_Tracker" customModuleProvider="target">
-                    <items>
-                        <group width="1" alignment="left" layout="vertical" spacing="12" id="xGm-4a-mJU">
-                            <items>
-                                <button width="1" alignment="left" title="Start" id="ba8-Ff-fOF">
-                                    <color key="titleColor" red="0.015686274509803921" green="0.87058823529411766" blue="0.44313725490196076" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
-                                    <connections>
-                                        <action selector="startMatch" destination="AgC-eL-Hgc" id="RVa-ai-hHJ"/>
-                                    </connections>
-                                </button>
-                                <button width="1" alignment="left" title="Stop" id="gNz-w0-UT6">
-                                    <color key="titleColor" red="0.84313725490000002" green="0.0" blue="0.070588235289999995" alpha="1" colorSpace="calibratedRGB"/>
-                                    <connections>
-                                        <action selector="stopMatch" destination="AgC-eL-Hgc" id="D6a-Nu-k2h"/>
-                                    </connections>
-                                </button>
-                                <timer alignment="center" id="gYX-Iq-gNi">
-                                    <color key="textColor" red="0.015686274509803921" green="0.87058823529411766" blue="0.44313725490196076" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
-                                    <fontDescription key="font" type="system" weight="semibold" pointSize="32"/>
-                                </timer>
-                            </items>
-                        </group>
-                    </items>
-                    <connections>
-                        <outlet property="timer" destination="gYX-Iq-gNi" id="i9m-j1-QZG"/>
-                    </connections>
-                </controller>
-            </objects>
-            <point key="canvasLocation" x="220" y="345"/>
-        </scene>
-        <!--Static Notification Interface Controller-->
-        <scene sceneID="AEw-b0-oYE">
-            <objects>
-                <notificationController id="YCC-NB-fut">
-                    <items>
-                        <label alignment="left" text="Alert Label" id="IdU-wH-bcW"/>
-                    </items>
-                    <notificationCategory key="notificationCategory" identifier="myCategory" id="JfB-70-Muf"/>
-                    <connections>
-                        <outlet property="notificationAlertLabel" destination="IdU-wH-bcW" id="JKC-fr-R95"/>
-                        <segue destination="4sK-HA-Art" kind="relationship" relationship="dynamicNotificationInterface" id="kXh-Jw-8B1"/>
-                    </connections>
-                </notificationController>
-            </objects>
-            <point key="canvasLocation" x="220" y="643"/>
-        </scene>
-        <!--Notification Controller-->
-        <scene sceneID="ZPc-GJ-vnh">
-            <objects>
-                <controller id="4sK-HA-Art" customClass="NotificationController" customModule="Match_Tracker" customModuleProvider="target"/>
-            </objects>
-            <point key="canvasLocation" x="468" y="643"/>
-        </scene>
-    </scenes>
-</document>
