+++ MatchTracker/Assets.xcassets/AppIcon.appiconset/Contents.json
