+++ Watch Extension/Assets.xcassets/Contents.json
